moved examples config file, added license where applicable, restructured unit tests
