+++ .gitignore
@@ -23,3 +23,4 @@ connectchain/config/config.yml
 .mcp.json
 config.yml
 .vscode/
+uv.lock
