+++ connectchain/lcel/model.py
@@ -44,31 +44,33 @@ def _get_model_(index: Any) -> BaseLanguageModel:
     model_config = models[index]
     if model_config is None:
         raise LCELModelException(f'Model config at index "{index}" is not defined')
-    
+
     # Check if we should use direct access (no EAS)
     needs_eas = False
     try:
         # Check if EAS is configured and not bypassed
-        if (hasattr(config, 'eas') and 
-            config.eas and 
-            hasattr(config.eas, 'id_key') and 
-            config.eas.id_key and
-            not getattr(model_config, 'bypass_eas', False)):
+        if (
+            hasattr(config, "eas")
+            and config.eas
+            and hasattr(config.eas, "id_key")
+            and config.eas.id_key
+            and not getattr(model_config, "bypass_eas", False)
+        ):
             needs_eas = True
     except (AttributeError, KeyError):
         needs_eas = False
-    
+
     # For non-OpenAI providers, always use direct access
     if model_config.provider != "openai":
         needs_eas = False
-    
+
     if needs_eas:
         # Use existing EAS flow for OpenAI
         model_instance = _get_openai_model_(index, config, model_config)
     else:
         # Use direct access for any provider
         model_instance = _get_direct_model_(model_config)
-    
+
     if model_instance is None:
         raise LCELModelException("Not implemented")
     try:
@@ -142,60 +144,60 @@ def _get_azure_model_(auth_token: str, model_config: Any) -> AzureOpenAI:
 
 def _get_direct_model_(model_config: Any) -> BaseLanguageModel:
     """Get a direct API model instance for any provider without EAS authentication"""
-    
+
     try:
         # Try using LangChain's automatic model initialization
         from langchain.chat_models import init_chat_model
-        
+
         # Prepare model name
         model_name = model_config.model_name
-        
+
         # Prepare configuration kwargs
         config_dict = {}
-        
+
         # Add custom API base if specified
-        if hasattr(model_config, 'api_base') and model_config.api_base:
-            config_dict['base_url'] = model_config.api_base
-        
+        if hasattr(model_config, "api_base") and model_config.api_base:
+            config_dict["base_url"] = model_config.api_base
+
         # Add temperature if specified
-        if hasattr(model_config, 'temperature'):
-            config_dict['temperature'] = model_config.temperature
-        
+        if hasattr(model_config, "temperature"):
+            config_dict["temperature"] = model_config.temperature
+
         # Handle custom API key environment variable
-        api_key_env = getattr(model_config, 'api_key_env', None)
+        api_key_env = getattr(model_config, "api_key_env", None)
         if api_key_env:
             api_key = os.getenv(api_key_env)
             if api_key:
-                config_dict['api_key'] = api_key
-        
+                config_dict["api_key"] = api_key
+
         # Try automatic initialization
         return init_chat_model(model_name, **config_dict)
-        
+
     except (ImportError, ValueError, Exception) as e:
         # Fallback to manual provider-specific initialization
         pass
-    
+
     # Manual provider-specific initialization as fallback
     # Determine API key environment variable
-    api_key_env = getattr(model_config, 'api_key_env', None)
+    api_key_env = getattr(model_config, "api_key_env", None)
     if not api_key_env:
         # Default to {PROVIDER}_API_KEY pattern
         api_key_env = f"{model_config.provider.upper()}_API_KEY"
-    
+
     api_key = os.getenv(api_key_env)
     if not api_key:
         raise LCELModelException(
             f"API key not found in environment variable: {api_key_env}. "
             f"Please set it in your .env file or environment."
         )
-    
+
     # Provider-specific instantiation
     if model_config.provider == "openai":
         # Check if this is Azure OpenAI based on api_base
-        api_base = getattr(model_config, 'api_base', None)
-        is_azure = api_base and 'openai.azure.com' in str(api_base)
-        
-        if is_azure and hasattr(model_config, 'api_version'):
+        api_base = getattr(model_config, "api_base", None)
+        is_azure = api_base and "openai.azure.com" in str(api_base)
+
+        if is_azure and hasattr(model_config, "api_version"):
             # Azure OpenAI with direct API key
             return AzureOpenAI(
                 model=model_config.model_name,
@@ -203,7 +205,7 @@ def _get_direct_model_(model_config: Any) -> BaseLanguageModel:
                 azure_endpoint=api_base,
                 api_version=model_config.api_version,
                 model_kwargs={
-                    "engine": getattr(model_config, 'engine', model_config.model_name),
+                    "engine": getattr(model_config, "engine", model_config.model_name),
                     "api_version": model_config.api_version,
                     "api_type": "azure",
                 },
@@ -215,23 +217,25 @@ def _get_direct_model_(model_config: Any) -> BaseLanguageModel:
                 api_key=api_key,
                 base_url=api_base,  # Can be None for default OpenAI endpoint
             )
-    
+
     elif model_config.provider == "anthropic":
         try:
             from langchain_anthropic import ChatAnthropic
+
             return ChatAnthropic(
                 model=model_config.model_name,
                 anthropic_api_key=api_key,
-                anthropic_api_url=getattr(model_config, 'api_base', None),
+                anthropic_api_url=getattr(model_config, "api_base", None),
             )
         except ImportError:
             raise LCELModelException(
                 "langchain-anthropic not installed. Run: pip install langchain-anthropic"
             )
-    
+
     elif model_config.provider == "google":
         try:
             from langchain_google_genai import ChatGoogleGenerativeAI
+
             return ChatGoogleGenerativeAI(
                 model=model_config.model_name,
                 google_api_key=api_key,
@@ -240,10 +244,11 @@ def _get_direct_model_(model_config: Any) -> BaseLanguageModel:
             raise LCELModelException(
                 "langchain-google-genai not installed. Run: pip install langchain-google-genai"
             )
-    
+
     elif model_config.provider == "cohere":
         try:
             from langchain_cohere import ChatCohere
+
             return ChatCohere(
                 model=model_config.model_name,
                 cohere_api_key=api_key,
@@ -252,20 +257,21 @@ def _get_direct_model_(model_config: Any) -> BaseLanguageModel:
             raise LCELModelException(
                 "langchain-cohere not installed. Run: pip install langchain-cohere"
             )
-    
+
     elif model_config.provider == "huggingface":
         try:
             from langchain_huggingface import HuggingFaceEndpoint
+
             return HuggingFaceEndpoint(
                 repo_id=model_config.model_name,
                 huggingfacehub_api_token=api_key,
-                endpoint_url=getattr(model_config, 'api_base', None),
+                endpoint_url=getattr(model_config, "api_base", None),
             )
         except ImportError:
             raise LCELModelException(
                 "langchain-huggingface not installed. Run: pip install langchain-huggingface"
             )
-    
+
     else:
         raise LCELModelException(
             f"Provider '{model_config.provider}' not supported. "
