+++ tests/unit_tests/tools/test_mcp.py
@@ -1,4 +1,16 @@
-"""Tests for MCPToolAgent."""
+# Copyright 2023 American Express Travel Related Services Company, Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
+# in compliance with the License. You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software distributed under the License
+# is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+# or implied. See the License for the specific language governing permissions and limitations under
+# the License.
+
+"""Tests for MCP tools."""
 
 from unittest.mock import AsyncMock, Mock, patch
 
@@ -11,6 +23,8 @@ from langchain.schema import AIMessage
 from langchain.tools import BaseTool
 
 from connectchain.tools.mcp.agent import MCPToolAgent
+from connectchain.tools.mcp.loader import MCPToolLoader
+from connectchain.utils.config import Config
 
 
 class TestMCPToolAgent:
@@ -123,3 +137,86 @@ class TestMCPToolAgent:
 
             with pytest.raises(RuntimeError, match="MCPToolAgent.invoke\\(\\) failed"):
                 agent.invoke({"query": "test"})
+
+
+class TestMCPToolLoader:
+    """Test cases for MCPToolLoader."""
+
+    @pytest.fixture
+    def mock_config(self):
+        """Create mock config with MCP settings."""
+        config = Mock(spec=Config)
+        config.data = {
+            "mcp": {
+                "servers": {
+                    "math_tools": {
+                        "command": "python",
+                        "args": ["math_server.py"],
+                        "transport": "stdio",
+                    },
+                    "web_tools": {
+                        "url": "https://example.com",
+                        "transport": "streamable-http",
+                        "auth": {"type": "bearer"},
+                    },
+                }
+            }
+        }
+        config.proxy = "http://proxy.company.com:8080"
+        config.eas = "https://eas.company.com"
+        config.cert = "/path/to/cert.pem"
+        return config
+
+    @pytest.mark.asyncio
+    async def test_load_all_tools(self, mock_config):
+        """Test loading tools from all configured servers."""
+        with patch("connectchain.tools.mcp.loader.MultiServerMCPClient") as mock_client_class:
+            mock_client = AsyncMock()
+            mock_client.get_tools = AsyncMock(
+                return_value=[Mock(name="add"), Mock(name="multiply")]
+            )
+            mock_client_class.return_value = mock_client
+
+            loader = MCPToolLoader(mock_config)
+            tools = await loader.load_tools()
+
+            assert len(tools) == 2
+            assert mock_client_class.called
+            assert mock_client.get_tools.called
+
+    @pytest.mark.asyncio
+    async def test_load_specific_tools(self, mock_config):
+        """Test loading tools from specific servers."""
+        with patch("connectchain.tools.mcp.loader.MultiServerMCPClient") as mock_client_class:
+            mock_client = AsyncMock()
+            mock_client.get_tools = AsyncMock(return_value=[Mock(name="add")])
+            mock_client_class.return_value = mock_client
+
+            loader = MCPToolLoader(mock_config)
+            await loader.load_tools(["math_tools"])
+
+            # Check that only math_tools was passed to client
+            call_args = mock_client_class.call_args[0][0]
+            assert "math_tools" in call_args
+            assert "web_tools" not in call_args
+
+
+    @pytest.mark.asyncio
+    async def test_no_servers_configured(self):
+        """Test behavior when no MCP servers are configured."""
+        config = Mock(spec=Config)
+        config.data = {}
+
+        loader = MCPToolLoader(config)
+        tools = await loader.load_tools()
+
+        assert tools == []
+
+    @pytest.mark.asyncio
+    async def test_close(self, mock_config):
+        """Test cleanup of MCP client."""
+        loader = MCPToolLoader(mock_config)
+        loader.client = Mock()
+
+        await loader.close()
+        assert loader.client is None
