--- tests/unit_tests/tools/test_mcp/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for MCP adapter integration."""
